+++ src/main/java/org/sopt/domain/member/exception/MemberValidationException.java
@@ -1,4 +1,4 @@
-package org.sopt.member.exception;
+package org.sopt.domain.member.exception;
 
 import org.sopt.global.exception.BaseException;
